+++ Packages/com.zebugger.zebug/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
+### Added
+ * **WIP** in-zebug window graph.
+   * E.G. `Zebug.GraphValue(Mathf.Sin(Time.time))`  
+
 ### Fix
 * Removed spurious profiling markers